trivial: remove a signed vs. unsigned warning on gcc

+++ client/pk-tools-common.c
@@ -152,7 +152,7 @@ pk_console_get_number (const gchar *question, guint maxnum)
 		retval = scanf("%u", &answer);
 
 		/* positive */
-		if (retval == 1 && answer > 0 && answer <= maxnum)
+		if (retval == 1 && answer > 0 && answer <= (gint) maxnum)
 			break;
 		g_print (_("Please enter a number from 1 to %i: "), maxnum);
 	} while (TRUE);
